requirements.txt
mcp[cli]>=1.5.0
google-generativeai>=0.7.0
httpx>=0.27.0
pyautogui>=0.9.54
numpy>=1.24.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0  # For vector operations
pillow>=10.0.0       # For image processing
